Refactor Architecture and Docker Configuration
