+++ backend/Dockerfile
@@ -4,7 +4,7 @@ COPY pom.xml .
 COPY src ./src
 RUN mvn clean package -DskipTests
 
-FROM eclipse-temurin:25-jdk-noble
+FROM eclipse-temurin:25-jre-noble
 WORKDIR /app
 COPY --from=build /app/target/*.jar app.jar
 EXPOSE 8081
